added SpiralPointGenerator
