+++ cs/TagCloud/CircularCloudLayouter.cs
@@ -8,10 +8,11 @@ public class CircularCloudLayouter : ICloudLayouter
     private readonly List<Rectangle> rectangles = new();
     private readonly IEnumerator<Point> spiralEnumerator;
 
-    public CircularCloudLayouter(Point center)
+    public CircularCloudLayouter(Point center, double spiralExpansionStep = 0.01, double spiralTwistStep = 0.01)
     {
         Center = center;
-        spiralEnumerator = GetSpiralPoints(Center, 0.01, 0.01).GetEnumerator();
+        spiralEnumerator = new SpiralPointGenerator(spiralExpansionStep, spiralTwistStep)
+            .Generate(center).GetEnumerator();
     }
 
     public Point Center { get; }
@@ -37,19 +38,5 @@ public class CircularCloudLayouter : ICloudLayouter
         var point = new Point(spiralEnumerator.Current.X - size.Width / 2,
             spiralEnumerator.Current.Y - size.Height / 2);
         return new Rectangle(point, size);
-    }
-
-    private static IEnumerable<Point> GetSpiralPoints(Point center, double dR, double dPhi)
-    {
-        double r = 0;
-        double phi = 0;
-        while (true)
-        {
-            r += dR;
-            phi += dPhi;
-            var x = (int)(r * Math.Cos(phi)) + center.X;
-            var y = (int)(r * Math.Sin(phi)) + center.Y;
-            yield return new Point(x, y);
-        }
     }
 }
