+++ cs/TagCloud/IPointGenerator.cs
@@ -1,0 +1,8 @@
+﻿using System.Drawing;
+
+namespace TagCloud;
+
+public interface IPointGenerator
+{
+    public IEnumerable<Point> Generate(Point center);
+}
